Module solana_runtime::accounts_background_service

Service to clean up dead slots in accounts_db

This can be expensive since we have to walk the append vecs being cleaned up.

Modules

stats 🔒

Stats for Accounts Background Services

Structs

AbsRequestHandlers

AbsRequestSender

AccountsBackgroundService

BANK_DROP_QUEUE_REPORTER 🔒

PrunedBankQueueLenReporter 🔒

PrunedBanksRequestHandler

SendDroppedBankCallback

SnapshotRequest

SnapshotRequestHandler

Enums

SnapshotRequestType

What type of request is this?

Constants

BANK_DROP_SIGNAL_CHANNEL_REPORT_INTERVAL 🔒

interval to report bank_drop queue events: 60s

CLEAN_INTERVAL_BLOCKS 🔒

INTERVAL_MS 🔒

MAX_DROP_BANK_SIGNAL_QUEUE_SIZE 🔒

maximum drop bank signal queue length

RECYCLE_STORE_EXPIRATION_INTERVAL_SECS 🔒

Functions

__stability 🔒

__static_ref_initialize 🔒

cmp_requests_by_priority 🔒

Compare snapshot requests; used to pick the highest priority request to handle.

new_accounts_package_type 🔒

Get the AccountsPackageType from a given SnapshotRequest

Type Definitions

DroppedSlotsReceiver

DroppedSlotsSender

SnapshotRequestReceiver

SnapshotRequestSender