Module solana_runtime::accounts_hash

Structs

AccountHashesFile

1 file containing account hashes sorted by pubkey

AccountsDeltaHash

Hash of accounts written in a single slot

AccountsHash

Hash of accounts

AccountsHasher

CalcAccountsHashConfig

parameters to calculate accounts hash

CalculateHashIntermediate

While scanning appendvecs, this is the info that needs to be extracted, de-duped, and sorted from what is stored in an append vec. Note this can be saved/loaded during hash calculation to a memory mapped file whose contents are CalculateHashIntermediate

CumulativeHashesFromFiles

used by merkle tree calculation to lookup account hashes by overall index

CumulativeOffset

CumulativeOffsets

HashStats

IncrementalAccountsHash

Hash of accounts that includes zero-lamport accounts Used with incremental snapshots

MmapAccountHashesFile 🔒

1 file containing account hashes sorted by pubkey, mapped into memory

Enums

AccountsHashEnum

Hash of accounts

ZeroLamportAccounts

How should zero-lamport accounts be treated by the accounts hasher?

Constants

MERKLE_FANOUT

THREE_LEVEL_OPTIMIZATION 🔒

Traits

ExtractSliceFromRawData

Type Definitions

SortedDataByPubkey

the data passed through the processing functions

StorageSizeQuartileStats