Module solana_runtime::accounts_index

Modules

ouroboros_impl_read_account_map_entry 🔒

Encapsulates implementation details for a self-referencing struct. This module is only visible when using –document-private-items.

Structs

AccountMapEntryInner

one entry in the in-mem accounts index Represents the value for an account key in the in-memory accounts index

AccountMapEntryMeta

data per entry in in-mem accounts index used to keep track of consistency with disk index

AccountSecondaryIndexes

AccountSecondaryIndexesIncludeExclude

AccountsIndex

T: account info type to interact in in-memory items U: account info type to be persisted to disk

AccountsIndexConfig

AccountsIndexIterator

AccountsIndexRootsStats

ReadAccountMapEntry

The self-referencing struct.

RootsTracker

ScanConfig

ScanSlotTracker

Enums

AccountIndex

AccountIndexGetResult

AccountsIndexScanResult

IndexKey

IndexLimitMb

specification of how much memory in-mem portion of account index can use

PreAllocatedAccountMapEntry

can be used to pre-allocate structures for insertion into accounts index outside of lock

ScanError

ScanTypes 🔒

UpsertReclaim

how accounts index ‘upsert’ should handle reclaims

Constants

ACCOUNTS_INDEX_CONFIG_FOR_BENCHMARKS

ACCOUNTS_INDEX_CONFIG_FOR_TESTING

BINS_DEFAULT

BINS_FOR_BENCHMARKS

BINS_FOR_TESTING

FLUSH_THREADS_TESTING

ITER_BATCH_SIZE

Statics

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

Traits

DiskIndexValue

IndexValue

IsCached

ZeroLamport

Type Definitions

AccountMap

AccountMapEntry 🔒

AccountMaps 🔒

LockMapType 🔒

LockMapTypeSlice 🔒

MapType 🔒

RefCount

ScanResult

SlotList

SlotSlice