Module solana_runtime::accounts_index::ouroboros_impl_read_account_map_entry

Encapsulates implementation details for a self-referencing struct. This module is only visible when using –document-private-items.

Structs

BorrowedFields 🔒

A struct for holding immutable references to all tail and immutably borrowed fields in an instance of ReadAccountMapEntry.

BorrowedMutFields 🔒

A struct for holding mutable references to all tail fields in an instance of ReadAccountMapEntry.

Heads 🔒

A struct which contains only the head fields of ReadAccountMapEntry.

ReadAccountMapEntry

The self-referencing struct.

ReadAccountMapEntryAsyncBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

ReadAccountMapEntryAsyncSendBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

ReadAccountMapEntryAsyncSendTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of makin your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

ReadAccountMapEntryAsyncTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of makin your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

ReadAccountMapEntryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

ReadAccountMapEntryTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of makin your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

Functions

check_if_okay_according_to_checkers 🔒

type_asserts 🔒