Struct solana_runtime::accounts_index::ouroboros_impl_read_account_map_entry::ReadAccountMapEntryBuilder

pub(super) struct ReadAccountMapEntryBuilder<T: IndexValue, SlotListGuardBuilder_: for<'this> FnOnce(&'this Arc<AccountMapEntryInner<T>>) -> RwLockReadGuard<'this, SlotList<T>>> {
    pub(super) owned_entry: Arc<AccountMapEntryInner<T>>,
    pub(super) slot_list_guard_builder: SlotListGuardBuilder_,
}

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

Field	Suggested Use

owned_entry	Directly pass in the value this field should contain

slot_list_guard_builder	Use a function or closure: (owned_entry: &_) -> slot_list_guard: _

Fields
owned_entry: Arc<AccountMapEntryInner<T>>slot_list_guard_builder: SlotListGuardBuilder_
Implementations

impl<T: IndexValue, SlotListGuardBuilder_: for<'this> FnOnce(&'this Arc<AccountMapEntryInner<T>>) -> RwLockReadGuard<'this, SlotList<T>>> ReadAccountMapEntryBuilder<T, SlotListGuardBuilder_>

pub(super) fn build(self) -> ReadAccountMapEntry<T>

Calls ReadAccountMapEntry::new() using the provided values. This is preferrable over calling new() directly for the reasons listed above.