Module solana_runtime::bank

The bank module tracks client accounts and the progress of on-chain programs.

A single bank relates to a block produced by a single leader and each bank except for the genesis bank points back to a parent bank.

The bank is the main entrypoint for processing verified transactions with the function Bank::process_transactions

It does this by loading the accounts using the reference it holds on the account store, and then passing those to an InvokeContext which handles loading the programs specified by the Transaction and executing it.

The bank then stores the results to the accounts store.

It then has APIs for retrieving if a transaction has been processed and it’s status. See get_signature_status et al.

Bank lifecycle:

A bank is newly created and open to transactions. Transactions are applied until either the bank reached the tick count when the node is the leader for that slot, or the node has applied all transactions present in all Entrys in the slot.

Once it is complete, the bank can then be frozen. After frozen, no more transactions can be applied or state changes made. At the frozen step, rent will be applied and various sysvar special accounts update to the new state of the system.

After frozen, and the bank has had the appropriate number of votes on it, then it can become rooted. At this point, it will not be able to be removed from the chain and the state is finalized.

It offers a high-level API that signs transactions on behalf of the caller, and a low-level API for when they have already been signed and verified.

Modules

address_lookup_table 🔒

builtin_programs 🔒

metrics 🔒

sysvar_cache 🔒

test_utils

utility function used for testing and benchmarking.

transaction_account_state_info 🔒

Structs

Bank

Manager for the state of all accounts and programs after processing its entries. AbiExample is needed even without Serialize/Deserialize; actual (de-)serialization are implemented elsewhere for versioning

BankFieldsToDeserialize

BankFieldsToSerialize 🔒

BankIncrementalSnapshotPersistence

Incremental snapshots only calculate their accounts hash based on the account changes WITHIN the incremental slot range. So, we need to keep track of the full snapshot expected accounts hash results. We also need to keep track of the hash and capitalization specific to the incremental snapshot slot range. The capitalization we calculate for the incremental slot will NOT be consistent with the bank’s capitalization. It is not feasible to calculate a capitalization delta that is correct given just incremental slots account data and the full snapshot’s capitalization.

BankRc

BankTestConfig

BuiltinPrograms

CollectRentFromAccountsInfo 🔒

Return the computed values from collect_rent_from_accounts()

CollectRentInPartitionInfo 🔒

Return the computed values—of each iteration in the parallel loop inside collect_rent_in_partition()—and then perform a reduce on all of them.

CommitTransactionCounts

InnerInstruction

LoadAndExecuteTransactionsOutput

LoadVoteAndStakeAccountsResult 🔒

NewBankOptions

NonceFull

Holds fee subtracted nonce info

NoncePartial

Holds limited nonce info available during transaction checks

OptionalDropCallback

PrevEpochInflationRewards 🔒

RentMetrics 🔒

RewardInfo

SquashTiming

StakeReward 🔒

TotalAccountsStats

Struct to collect stats when scanning all accounts in get_total_accounts_stats()

TransactionBalancesSet

TransactionExecutionDetails

TransactionLogCollector

TransactionLogCollectorConfig

TransactionLogInfo

TransactionResults

TransactionSimulationResult

VerifyAccountsHashConfig

params to verify_accounts_hash

VoteWithStakeDelegations 🔒

__FieldVisitor 🔒

__FieldVisitor 🔒

__FieldVisitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

Enums

ApplyFeatureActivationsCaller 🔒

Since apply_feature_activations() has different behavior depending on its caller, enumerate those callers explicitly.

DurableNonceFee

RewardCalculationEvent

RewardType

TransactionExecutionResult

Type safe representation of a transaction execution attempt which differentiates between a transaction that was executed (will be committed to the ledger) and a transaction which wasn’t executed and will be dropped.

TransactionLogCollectorFilter

__Field 🔒

__Field 🔒

__Field 🔒

Constants

ACCOUNTS_DATA_LEN 🔒

ACCOUNT_DATA_COST_PAGE_SIZE 🔒

BASE_CONGESTION 🔒

EVICT_CACHE_TO_PERCENTAGE 🔒

FIELDS 🔒

FIELDS 🔒

MAX_LEADER_SCHEDULE_STAKES

NUM_BLOCKHASH_CONFIRMATIONS 🔒

PREFIX_MAX 🔒

PREFIX_MAX 🔒

PREFIX_SIZE 🔒

PREFIX_SIZE 🔒

RENT_UNADJUSTED_INITIAL_BALANCE 🔒

SECONDS_PER_YEAR

VARIANTS 🔒

Statics

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INC_NEW_COUNTER 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

INIT_HOOK 🔒

Traits

DropCallback

NonceInfo

RewardCalcTracer

type alias is not supported for trait in rust yet. As a workaround, we define the RewardCalcTracer trait explicitly and implement it on any type that implement Fn(&RewardCalculationEvent) + Send + Sync.

Functions

calculate_data_size_delta 🔒

Compute how much an account has changed size. This function is useful when the data size delta needs to be computed and passed to an update_accounts_data_size_delta function.

inner_instructions_list_from_instruction_trace

Extract the InnerInstructionsList from a TransactionContext

merge 🔒

new 🔒

null_tracer 🔒

Type Definitions

BankSlotDelta

BankStatusCache

EpochCount 🔒

InnerInstructions

An ordered list of compiled instructions that were invoked during a transaction instruction

InnerInstructionsList

A list of compiled instructions that were invoked during each instruction of a transaction

InvalidCacheKeyMap 🔒

Partition 🔒

PartitionIndex 🔒

PartitionsPerCycle

Prefix 🔒

Prefix 🔒

RentCollectionCycleParams 🔒

StakeRewards 🔒

TransactionBalances

TransactionCheckResult

TransactionLogMessages

A list of log messages emitted during a transaction

VoteRewards 🔒

VoteWithStakeDelegationsMap 🔒