Struct solana_runtime::bucket_map_holder_stats::BucketMapHolderStats

pub struct BucketMapHolderStats {

    pub held_in_mem: BucketMapHeldInMemStats,
    pub get_mem_us: AtomicU64,
    pub gets_from_mem: AtomicU64,
    pub get_missing_us: AtomicU64,
    pub gets_missing: AtomicU64,
    pub entry_mem_us: AtomicU64,
    pub entries_from_mem: AtomicU64,
    pub entry_missing_us: AtomicU64,
    pub entries_missing: AtomicU64,
    pub load_disk_found_count: AtomicU64,
    pub load_disk_found_us: AtomicU64,
    pub load_disk_missing_count: AtomicU64,
    pub load_disk_missing_us: AtomicU64,
    pub updates_in_mem: AtomicU64,
    pub items: AtomicU64,
    pub items_us: AtomicU64,
    pub failed_to_evict: AtomicU64,
    pub keys: AtomicU64,
    pub deletes: AtomicU64,
    pub buckets_scanned: AtomicU64,
    pub inserts: AtomicU64,
    count: AtomicUsize,
    pub bg_waiting_us: AtomicU64,
    pub bg_throttling_wait_us: AtomicU64,
    pub count_in_mem: AtomicUsize,
    pub per_bucket_count: Vec<AtomicUsize>,
    pub flush_entries_updated_on_disk: AtomicU64,
    pub flush_entries_evicted_from_mem: AtomicU64,
    pub active_threads: AtomicU64,
    pub get_range_us: AtomicU64,
    last_age: AtomicU8,
    last_ages_flushed: AtomicU64,
    pub flush_scan_us: AtomicU64,
    pub flush_update_us: AtomicU64,
    pub flush_evict_us: AtomicU64,
    pub flush_grow_us: AtomicU64,
    last_was_startup: AtomicBool,
    last_time: AtomicInterval,
    bins: u64,
    pub estimate_mem: AtomicU64,
    pub flush_should_evict_us: AtomicU64,

}

Fields
held_in_mem: BucketMapHeldInMemStatsget_mem_us: AtomicU64gets_from_mem: AtomicU64get_missing_us: AtomicU64gets_missing: AtomicU64entry_mem_us: AtomicU64entries_from_mem: AtomicU64entry_missing_us: AtomicU64entries_missing: AtomicU64load_disk_found_count: AtomicU64load_disk_found_us: AtomicU64load_disk_missing_count: AtomicU64load_disk_missing_us: AtomicU64updates_in_mem: AtomicU64items: AtomicU64items_us: AtomicU64failed_to_evict: AtomicU64keys: AtomicU64deletes: AtomicU64buckets_scanned: AtomicU64inserts: AtomicU64count: AtomicUsizebg_waiting_us: AtomicU64bg_throttling_wait_us: AtomicU64count_in_mem: AtomicUsizeper_bucket_count: Vec<AtomicUsize>flush_entries_updated_on_disk: AtomicU64flush_entries_evicted_from_mem: AtomicU64active_threads: AtomicU64get_range_us: AtomicU64last_age: AtomicU8last_ages_flushed: AtomicU64flush_scan_us: AtomicU64flush_update_us: AtomicU64flush_evict_us: AtomicU64flush_grow_us: AtomicU64last_was_startup: AtomicBoollast_time: AtomicIntervalbins: u64estimate_mem: AtomicU64flush_should_evict_us: AtomicU64
Implementations

impl BucketMapHolderStats

pub fn new(bins: usize) -> BucketMapHolderStats

pub fn inc_insert(&self)

pub fn inc_insert_count(&self, count: u64)

pub fn inc_delete(&self)

pub fn inc_mem_count(&self, bin: usize)

pub fn dec_mem_count(&self, bin: usize)

pub fn add_mem_count(&self, bin: usize, count: usize)

pub fn sub_mem_count(&self, bin: usize, count: usize)

fn ms_per_age<T: IndexValue, U: DiskIndexValue + From<T> + Into<T>>( &self, storage: &BucketMapHolder<T, U>, elapsed_ms: u64 ) -> u64

pub fn remaining_until_next_interval(&self) -> u64

fn get_stats(data: Vec<usize>) -> (usize, usize, usize, usize)

return min, max, sum, median of data

fn calc_percent(ms: u64, elapsed_ms: u64) -> f32

pub fn total_count(&self) -> usize

pub fn count_in_bucket(&self, bucket: usize) -> usize

pub fn get_remaining_items_to_flush_estimate(&self) -> usize

This is an estimate of the # of items in mem that are awaiting flushing to disk. returns (# items in mem) - (# items we intend to hold in mem for performance heuristics) The result is also an estimate because ‘held_in_mem’ is based on a stat that is swapped out when stats are reported.

pub fn report_stats<T: IndexValue, U: DiskIndexValue + From<T> + Into<T>>( &self, storage: &BucketMapHolder<T, U> )