Struct solana_runtime::message_processor::MessageProcessor

pub struct MessageProcessor {}

Implementations

impl MessageProcessor

pub fn process_message( builtin_programs: &[BuiltinProgram], message: &SanitizedMessage, program_indices: &[Vec<IndexOfAccount>], transaction_context: &mut TransactionContext, rent: Rent, log_collector: Option<Rc<RefCell<LogCollector>>>, tx_executor_cache: Rc<RefCell<TransactionExecutorCache>>, feature_set: Arc<FeatureSet>, compute_budget: ComputeBudget, timings: &mut ExecuteTimings, sysvar_cache: &SysvarCache, blockhash: Hash, lamports_per_signature: u64, current_accounts_data_len: u64, accumulated_consumed_units: &mut u64 ) -> Result<ProcessedMessageInfo, TransactionError>

Process a message. This method calls each instruction in the message over the set of loaded accounts. For each instruction it calls the program entrypoint method and verifies that the result of the call does not violate the bank’s accounting rules. The accounts are committed back to the bank only if every instruction succeeds.