Struct solana_runtime::prioritization_fee::PrioritizationFee

pub struct PrioritizationFee {
    min_transaction_fee: u64,
    min_writable_account_fees: HashMap<Pubkey, u64>,
    is_finalized: bool,
    metrics: PrioritizationFeeMetrics,
}

Block minimum prioritization fee stats, includes the minimum prioritization fee for a transaction in this block; and the minimum fee for each writable account in all transactions in this block. The only relevant write account minimum fees are those greater than the block minimum transaction fee, because the minimum fee needed to land a transaction is determined by Max( min_transaction_fee, min_writable_account_fees(key), …)

Fields
min_transaction_fee: u64min_writable_account_fees: HashMap<Pubkey, u64>is_finalized: boolmetrics: PrioritizationFeeMetrics
Implementations

impl PrioritizationFee

pub fn update( &mut self, transaction_fee: u64, writable_accounts: &[Pubkey] ) -> Result<(), PrioritizationFeeError>

Update self for minimum transaction fee in the block and minimum fee for each writable account.

fn prune_irrelevant_writable_accounts(&mut self)

Accounts that have minimum fees lesser or equal to the minimum fee in the block are redundant, they are removed to reduce memory footprint when mark_block_completed() is called.

pub fn mark_block_completed(&mut self) -> Result<(), PrioritizationFeeError>

pub fn get_min_transaction_fee(&self) -> Option<u64>

pub fn get_writable_account_fee(&self, key: &Pubkey) -> Option<u64>

pub fn get_writable_account_fees(&self) -> impl Iterator<Item = (&Pubkey, &u64)>

pub fn get_writable_accounts_count(&self) -> usize

pub fn is_finalized(&self) -> bool

pub fn report_metrics(&self, slot: Slot)