Struct solana_runtime::rolling_bit_field::RollingBitField

pub struct RollingBitField {
    max_width: u64,
    min: u64,
    max_exclusive: u64,
    bits: BitVec,
    count: usize,
    excess: HashSet<u64>,
}

Fields
max_width: u64min: u64max_exclusive: u64bits: BitVeccount: usizeexcess: HashSet<u64>
Implementations

impl RollingBitField

functionally similar to a hashset Relies on there being a sliding window of key values. The key values continue to increase. Old key values are removed from the lesser values and do not accumulate.

pub fn new(max_width: u64) -> Self

fn get_address(&self, key: &u64) -> u64

pub fn range_width(&self) -> u64

pub fn min(&self) -> Option<u64>

pub fn insert(&mut self, key: u64)

pub fn remove(&mut self, key: &u64) -> bool

remove key from set, return if item was in the set

fn all_items_in_excess(&self) -> bool

fn purge(&mut self, key: &u64)

fn contains_assume_in_range(&self, key: &u64) -> bool

pub fn contains(&self, key: &u64) -> bool

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn max_exclusive(&self) -> u64

pub fn max_inclusive(&self) -> u64

pub fn get_all_less_than(&self, max_slot_exclusive: Slot) -> Vec<u64>

return all items < ‘max_slot_exclusive’

pub fn get_prior(&self, max_slot_exclusive: Slot) -> Option<Slot>

return highest item < ‘max_slot_exclusive’

pub fn get_all(&self) -> Vec<u64>