Module solana_runtime::snapshot_utils::snapshot_storage_rebuilder

Provides interfaces for rebuilding snapshot storages

Structs

BANK_FIELDS_FILE_REGEX ğŸ”’

RebuiltSnapshotStorage ğŸ”’

Convenient wrapper for snapshot version and rebuilt storages

STORAGE_FILE_REGEX ğŸ”’

SnapshotStorageRebuilder ğŸ”’

Stores state for rebuilding snapshot storages

VERSION_FILE_REGEX ğŸ”’

Enums

SnapshotFileKind ğŸ”’

Used to determine if a filename is structured like a version file, bank file, or storage file

Functions

__stability ğŸ”’

__stability ğŸ”’

__stability ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

get_slot_and_append_vec_id ğŸ”’

Get the slot and append vec id from the filename

get_snapshot_file_kind ğŸ”’

Determines SnapshotFileKind for filename if any