Struct solana_runtime::stakes::Stakes

pub struct Stakes<T: Clone> {
    vote_accounts: VoteAccounts,
    stake_delegations: ImHashMap<Pubkey, T>,
    unused: u64,
    epoch: Epoch,
    stake_history: StakeHistory,
}

The generic type T is either Delegation or StakeAccount. Stakes<Delegation> is equivalent to the old code and is used for backward compatibility in crate::bank::BankFieldsToDeserialize. But banks cache Stakes<StakeAccount> which includes the entire stake account and StakeState deserialized from the account. Doing so, will remove the need to load the stake account from accounts-db when working with stake-delegations.

Fields
vote_accounts: VoteAccounts

vote accounts

stake_delegations: ImHashMap<Pubkey, T>

stake_delegations

unused: u64

unused

epoch: Epoch

current epoch, used to calculate current stake

stake_history: StakeHistory

history of staking levels

Implementations

impl<T: Clone> Stakes<T>

pub fn vote_accounts(&self) -> &VoteAccounts

pub(crate) fn staked_nodes(&self) -> Arc<HashMap<Pubkey, u64>>

impl Stakes<StakeAccount<Delegation>>

pub(crate) fn new<F>( stakes: &Stakes<Delegation>, get_account: F ) -> Result<Self, Error>where F: Fn(&Pubkey) -> Option<AccountSharedData>,

Creates a Stake from Stake by loading the full account state for respective stake pubkeys. get_account function should return the account at the respective slot where stakes where cached.

pub(crate) fn history(&self) -> &StakeHistory

fn activate_epoch(&mut self, next_epoch: Epoch, thread_pool: &ThreadPool)

fn calculate_stake( &self, voter_pubkey: &Pubkey, epoch: Epoch, stake_history: &StakeHistory ) -> u64

Sum the stakes that point to the given voter_pubkey

pub(crate) fn vote_balance_and_staked(&self) -> u64

Sum the lamports of the vote accounts and the delegated stake

fn remove_vote_account(&mut self, vote_pubkey: &Pubkey)

fn remove_stake_delegation(&mut self, stake_pubkey: &Pubkey)

fn upsert_vote_account( &mut self, vote_pubkey: &Pubkey, vote_account: VoteAccount )

fn upsert_stake_delegation( &mut self, stake_pubkey: Pubkey, stake_account: StakeAccount<Delegation> )

pub(crate) fn stake_delegations( &self ) -> &ImHashMap<Pubkey, StakeAccount<Delegation>>

pub(crate) fn highest_staked_node(&self) -> Option<Pubkey>