Trait solana_runtime::storable_accounts::StorableAccounts

pub trait StorableAccounts<'a, T: ReadableAccount + Sync>: Sync {
    // Required methods
    fn pubkey(&self, index: usize) -> &Pubkey;
    fn account(&self, index: usize) -> &T;
    fn slot(&self, index: usize) -> Slot;
    fn target_slot(&self) -> Slot;
    fn len(&self) -> usize;
    fn include_slot_in_hash(&self) -> IncludeSlotInHash;

    // Provided methods
    fn account_default_if_zero_lamport(&self, index: usize) -> Option<&T> { ... }
    fn is_empty(&self) -> bool { ... }
    fn contains_multiple_slots(&self) -> bool { ... }
    fn has_hash_and_write_version(&self) -> bool { ... }
    fn hash(&self, _index: usize) -> &Hash { ... }
    fn write_version(&self, _index: usize) -> u64 { ... }
}

abstract access to pubkey, account, slot, target_slot of either: a. (slot, &[&Pubkey, &ReadableAccount]) b. (slot, &[&Pubkey, &ReadableAccount, Slot]) (we will use this later) This trait avoids having to allocate redundant data when there is a duplicated slot parameter. All legacy callers do not have a unique slot per account to store.

Required Methods

fn pubkey(&self, index: usize) -> &Pubkey

pubkey at ‘index’

fn account(&self, index: usize) -> &T

account at ‘index’

fn slot(&self, index: usize) -> Slot

fn target_slot(&self) -> Slot

slot that all accounts are to be written to

fn len(&self) -> usize

accounts to write

fn include_slot_in_hash(&self) -> IncludeSlotInHash

true iff hashing these accounts should include the slot

Provided Methods

fn account_default_if_zero_lamport(&self, index: usize) -> Option<&T>

None if account is zero lamports

fn is_empty(&self) -> bool

true if no accounts to write

fn contains_multiple_slots(&self) -> bool

are there accounts from multiple slots only used for an assert

fn has_hash_and_write_version(&self) -> bool

true iff the impl can provide hash and write_version Otherwise, hash and write_version have to be provided separately to store functions.

fn hash(&self, _index: usize) -> &Hash

return hash for account at ‘index’ Should only be called if ‘has_hash_and_write_version’ = true

fn write_version(&self, _index: usize) -> u64

return write_version for account at ‘index’ Should only be called if ‘has_hash_and_write_version’ = true