Crate solana_runtime

Modules

account_info

AccountInfo represents a reference to AccountSharedData in either an AppendVec or the write cache. AccountInfo is not persisted anywhere between program runs. AccountInfo is purely runtime state. Note that AccountInfo is saved to disk buckets during runtime, but disk buckets are recreated at startup.

account_overrides

account_rent_state

account_storage

Manage the map of slot -> append vec

accounts

accounts_background_service

Service to clean up dead slots in accounts_db

accounts_cache

accounts_db

Persistent accounts are stored at this path location: <path>/<pid>/data/

accounts_file

accounts_hash

accounts_index

accounts_index_storage

accounts_update_notifier_interface

active_stats 🔒

keep track of areas of the validator that are currently active

ancestors

ancient_append_vecs 🔒

helpers for squashing append vecs into ancient append vecs an ancient append vec is:

append_vec

Persistent storage for accounts.

bank

The bank module tracks client accounts and the progress of on-chain programs.

bank_client

bank_creation_freezing_progress 🔒

Keep track of how many banks have been created and how many have been frozen or dropped. This is useful to track foreground progress to understand expected access to accounts db.

bank_forks

The bank_forks module implements BankForks a DAG of checkpointed Banks

bank_utils

block_cost_limits

defines block cost related limits

blockhash_queue

bucket_map_holder

bucket_map_holder_stats

builtins

cache_hash_data

Cached data for hashing accounts

cache_hash_data_stats

Cached data for hashing accounts

commitment

contains

cost_model

’cost_model` provides service to estimate a transaction’s cost following proposed fee schedule #16984; Relevant cluster cost measuring is described by #19627

cost_tracker

cost_tracker keeps tracking transaction cost per chained accounts as well as for entire block The main functions are:

epoch_accounts_hash

The Epoch Accounts Hash (EAH) is a special hash of the whole accounts state that occurs once per epoch.

epoch_stakes

genesis_utils

hardened_unpack

in_mem_accounts_index

inline_spl_associated_token_account

inline_spl_token

inline_spl_token_2022

installed_scheduler_pool

Transaction processing glue code, mainly consisting of Object-safe traits

loader_utils

message_processor

non_circulating_supply

nonce_keyed_account 🔒

prioritization_fee

prioritization_fee_cache

pubkey_bins 🔒

read_only_accounts_cache 🔒

ReadOnlyAccountsCache used to store accounts, such as executable accounts, which can be large, loaded many times, and rarely change.

rent_collector

calculate and collect rent from Accounts

rent_debits

rent_paying_accounts_by_partition 🔒

Provide fast iteration of all pubkeys which could possibly be rent paying, grouped by rent collection partition

rolling_bit_field 🔒

functionally similar to a hashset Relies on there being a sliding window of key values. The key values continue to increase. Old key values are removed from the lesser values and do not accumulate.

root_bank_cache

A wrapper around a root Bank that only loads from bank forks if the root has been updated. This can be useful to avoid read-locking the bank forks until the root has been updated.

runtime_config

secondary_index

serde_snapshot

shared_buffer_reader 🔒

SharedBuffer is given a Reader and SharedBufferReader implements the Reader trait. SharedBuffer reads ahead in the underlying file and saves the data. SharedBufferReaders can be created for the buffer and independently keep track of each reader’s read location. The background reader keeps track of the progress of each client. After data has been read by all readers, the buffer is recycled and reading ahead continues. A primary use case is the underlying reader being decompressing a file, which can be computationally expensive. The clients of SharedBufferReaders could be parallel instances which need access to the decompressed data.

snapshot_archive_info

Information about snapshot archives

snapshot_config

snapshot_hash

Helper types and functions for handling and dealing with snapshot hashes.

snapshot_minimizer

Used to create minimal snapshots - separated here to keep accounts_db simpler

snapshot_package

snapshot_utils

sorted_storages

stake_account 🔒

stake_history

This module implements clone-on-write semantics for the SDK’s StakeHistory to reduce unnecessary cloning of the underlying vector.

stake_weighted_timestamp

stakes

Stakes serve as a cache of stake and vote accounts to derive node stakes

static_ids

status_cache

storable_accounts 🔒

trait for abstracting underlying storage of pubkey and account pairs to be written

system_instruction_processor 🔒

transaction_batch

transaction_error_metrics

transaction_priority_details

verify_accounts_hash_in_background 🔒

at startup, verify accounts hash in the background

vote_account

vote_parser

vote_sender_types

vote_transaction

waitable_condvar

Macros

u64_align