Module solana_runtime::append_vec

Persistent storage for accounts.

For more information, see:

https://docs.solana.com/implemented-proposals/persistent-account-storage

Modules

test_utils

Helpers for AppendVec tests and benches

Structs

APPEND_VEC_MMAPPED_FILES_OPEN

AppendVec

A thread-safe, file-backed block of memory used to store Account instances. Append operations are serialized such that only one thread updates the internal append_lock at a time. No restrictions are placed on reading. That is, one may read items from one thread while another is appending new items.

AppendVecAccountsIter

AppendVecStoredAccountMeta

References to account data stored elsewhere. Getting an Account requires cloning (see StoredAccountMeta::clone_account()).

Enums

MatchAccountOwnerError

Constants

MAXIMUM_APPEND_VEC_FILE_SIZE

STORE_META_OVERHEAD

size of the fixed sized fields in an append vec we need to add data len and align it to get the actual stored size

Statics

INC_NEW_COUNTER 🔒

INIT_HOOK 🔒

Functions

__stability 🔒

__static_ref_initialize 🔒

aligned_stored_size

Returns the size this item will take to store plus possible alignment padding bytes before the next entry. fixed-size portion of per-account data written plus ‘data_len’, aligned to next boundary