Module solana_runtime::snapshot_package

Modules

compare 🔒

Structs

AccountsPackage

This struct packages up fields to send from AccountsBackgroundService to AccountsHashVerifier

SnapshotPackage

This struct packages up fields to send from AccountsHashVerifier to SnapshotPackagerService

SupplementalSnapshotInfo

Supplemental information needed for snapshots

Enums

AccountsPackageType

Accounts packages are sent to the Accounts Hash Verifier for processing. There are multiple types of accounts packages, which are specified as variants in this enum. All accounts packages do share some processing: such as calculating the accounts hash.

SnapshotType

Snapshots come in two flavors, Full and Incremental. The IncrementalSnapshot has a Slot field, which is the incremental snapshot base slot.

Functions

cmp_accounts_package_types_by_priority

Compare accounts package types by priority

cmp_accounts_packages_by_priority

Compare accounts packages by priority; first by type, then by slot

cmp_snapshot_packages_by_priority

Compare snapshot packages by priority; first by type, then by slot

cmp_snapshot_types_by_priority

Compare snapshot types by priority

retain_max_n_elements

Helper function to retain only max n of elements to the right of a vector, viz. remove v.len() - n elements from the left of the vector.